pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dndml"
version = "0.1.0"
description = "A simple tool for creating and managing D&D 5e resources."
authors = ["The Aufish <the@aufish.com>"]
readme = "README.md"
repository = "https://github.com/theaufish-git/dndml"
packages = [
    { include = "dndml", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
poetry = "^1.8.2"
pydantic = "^2.6.4"
make = "^0.1.6.post2"
lint = "^1.2.1"
case-converter = "^1.1.0"
pyyaml = "^6.0.1"
